- switched from Picasso to Glide - fixed database schemas - adding items to cart is now supported

+++ app/src/main/java/com/example/store/features/dashboard/ui/CategoryAdapter.kt
@@ -22,7 +22,7 @@ private object CategoryRecyclerAdapterCallback : DiffUtil.ItemCallback<CategoryV
 }
 
 class CategoryAdapter (
-    private val itemListener: ()->Unit
+    private val itemListener: (Int)->Unit
 ):
     ListAdapter<CategoryView, CategoryAdapter.ViewHolder>(CategoryRecyclerAdapterCallback) {
     private lateinit var context: Context
